Add controller, service and entity classes for Department and Student
Implemented controller classes for Department and Student. Additionally, classes for Department and Student were added to the dao, entity, and service packages. Also, added corresponding test files and necessary resources such as application.yml and SQL files. These additions are a part of setting up a RESTful API for the simple student-management system.